+++ common/RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # dxCommon
 
+## in develop
+
+* Adds optional `localizationDir` parameter to `LocalizationDisambiguator` methods for specifying the localization directory that must be used
+
 ## 0.4.1 (2021-06-08)
 
 * Fixes issues with local and http path relativization